fix : 친구요청 api

+++ frontend/ARchive/app/src/main/java/com/droidblossom/archive/domain/model/friend/FriendAcceptRequest.kt
@@ -4,7 +4,7 @@ import com.droidblossom.archive.data.dto.friend.request.FriendAcceptRequestDto
 
 
 data class FriendAcceptRequest (
-    val friendId : Double
+    val friendId : Long
 ){
     fun toDto() = FriendAcceptRequestDto(
         friendId = this.friendId,
